notification-service/requirements.txt
# ============================================
# Notification Service - Dependencies
# ============================================

# Async HTTP client
httpx==0.25.2

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
hiredis==2.2.3

# Async framework
asyncio==3.4.3

# Logging
loguru==0.7.2

# Configuration
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Utilities
python-dateutil==2.8.2